third_party/blink/web_tests/external/wpt/html/rendering/replaced-elements/images/img-sizes-auto-expected.txt

This is a testharness.js-based test.
[FAIL] <img sizes="auto"> - contain
  assert_equals: expected "size" but got "none"
[FAIL] <img sizes="auto"> - contain-intrinsic-size
  assert_equals: expected "300px 150px" but got "none"
[FAIL] <img sizes="auto" width="10" height="20"> - contain
  assert_equals: expected "size" but got "none"
[FAIL] <img sizes="auto" width="10" height="20"> - contain-intrinsic-size
  assert_equals: expected "300px 150px" but got "none"
[FAIL] <img sizes="AuTo"> - contain
  assert_equals: expected "size" but got "none"
[FAIL] <img sizes="AuTo"> - contain-intrinsic-size
  assert_equals: expected "300px 150px" but got "none"
[FAIL] <img sizes="auto,xyz"> - contain
  assert_equals: expected "size" but got "none"
[FAIL] <img sizes="auto,xyz"> - contain-intrinsic-size
  assert_equals: expected "300px 150px" but got "none"
[FAIL] <img sizes="AuTo,xyz"> - contain
  assert_equals: expected "size" but got "none"
[FAIL] <img sizes="AuTo,xyz"> - contain-intrinsic-size
  assert_equals: expected "300px 150px" but got "none"
[FAIL] <img sizes="auto" class="test-important"> - contain
  assert_equals: expected "size" but got "style"
[FAIL] <img sizes="auto" class="test-important"> - contain-intrinsic-size
  assert_equals: expected "300px 150px" but got "30px 15px"
Harness: the test ran to completion.
